Moving multiple classes out of RecordOperations.php to allow for autoloading to work consistently.
