+++ src/com/zoho/crm/api/record/GetRecordHeader.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace com\zoho\crm\api\record;
+
+use com\zoho\crm\api\Header;
+
+class GetRecordHeader
+{
+    public static final function IfModifiedSince()
+    {
+        return new Header('If-Modified-Since', 'com.zoho.crm.api.Record.GetRecordHeader');
+    }
+}
