+++ src/com/zoho/crm/api/record/RecordOperations.php
@@ -391,258 +391,4 @@ class RecordOperations
 		return $handlerInstance->apiCall(MassUpdateResponseHandler::class, 'application/json'); 
 
 	}
-} 
-class GetRecordParam
-{
-	public static final function approved()
-	{
-		return new Param('approved', 'com.zoho.crm.api.Record.GetRecordParam'); 
-
-	}
-	public static final function converted()
-	{
-		return new Param('converted', 'com.zoho.crm.api.Record.GetRecordParam'); 
-
-	}
-	public static final function cvid()
-	{
-		return new Param('cvid', 'com.zoho.crm.api.Record.GetRecordParam'); 
-
-	}
-	public static final function uid()
-	{
-		return new Param('uid', 'com.zoho.crm.api.Record.GetRecordParam'); 
-
-	}
-	public static final function fields()
-	{
-		return new Param('fields', 'com.zoho.crm.api.Record.GetRecordParam'); 
-
-	}
-	public static final function startDateTime()
-	{
-		return new Param('startDateTime', 'com.zoho.crm.api.Record.GetRecordParam'); 
-
-	}
-	public static final function endDateTime()
-	{
-		return new Param('endDateTime', 'com.zoho.crm.api.Record.GetRecordParam'); 
-
-	}
-	public static final function territoryId()
-	{
-		return new Param('territory_id', 'com.zoho.crm.api.Record.GetRecordParam'); 
-
-	}
-	public static final function includeChild()
-	{
-		return new Param('include_child', 'com.zoho.crm.api.Record.GetRecordParam'); 
-
-	}
-
-}
-
-class GetRecordHeader
-{
-	public static final function IfModifiedSince()
-	{
-		return new Header('If-Modified-Since', 'com.zoho.crm.api.Record.GetRecordHeader'); 
-
-	}
-
-}
-
-class DeleteRecordParam
-{
-	public static final function wfTrigger()
-	{
-		return new Param('wf_trigger', 'com.zoho.crm.api.Record.DeleteRecordParam'); 
-
-	}
-
-}
-
-class GetRecordsParam
-{
-	public static final function approved()
-	{
-		return new Param('approved', 'com.zoho.crm.api.Record.GetRecordsParam'); 
-
-	}
-	public static final function converted()
-	{
-		return new Param('converted', 'com.zoho.crm.api.Record.GetRecordsParam'); 
-
-	}
-	public static final function cvid()
-	{
-		return new Param('cvid', 'com.zoho.crm.api.Record.GetRecordsParam'); 
-
-	}
-	public static final function ids()
-	{
-		return new Param('ids', 'com.zoho.crm.api.Record.GetRecordsParam'); 
-
-	}
-	public static final function uid()
-	{
-		return new Param('uid', 'com.zoho.crm.api.Record.GetRecordsParam'); 
-
-	}
-	public static final function fields()
-	{
-		return new Param('fields', 'com.zoho.crm.api.Record.GetRecordsParam'); 
-
-	}
-	public static final function sortBy()
-	{
-		return new Param('sort_by', 'com.zoho.crm.api.Record.GetRecordsParam'); 
-
-	}
-	public static final function sortOrder()
-	{
-		return new Param('sort_order', 'com.zoho.crm.api.Record.GetRecordsParam'); 
-
-	}
-	public static final function page()
-	{
-		return new Param('page', 'com.zoho.crm.api.Record.GetRecordsParam'); 
-
-	}
-	public static final function perPage()
-	{
-		return new Param('per_page', 'com.zoho.crm.api.Record.GetRecordsParam'); 
-
-	}
-	public static final function startDateTime()
-	{
-		return new Param('startDateTime', 'com.zoho.crm.api.Record.GetRecordsParam'); 
-
-	}
-	public static final function endDateTime()
-	{
-		return new Param('endDateTime', 'com.zoho.crm.api.Record.GetRecordsParam'); 
-
-	}
-	public static final function territoryId()
-	{
-		return new Param('territory_id', 'com.zoho.crm.api.Record.GetRecordsParam'); 
-
-	}
-	public static final function includeChild()
-	{
-		return new Param('include_child', 'com.zoho.crm.api.Record.GetRecordsParam'); 
-
-	}
-
-}
-
-class GetRecordsHeader
-{
-	public static final function IfModifiedSince()
-	{
-		return new Header('If-Modified-Since', 'com.zoho.crm.api.Record.GetRecordsHeader'); 
-
-	}
-
-}
-
-class DeleteRecordsParam
-{
-	public static final function ids()
-	{
-		return new Param('ids', 'com.zoho.crm.api.Record.DeleteRecordsParam'); 
-
-	}
-	public static final function wfTrigger()
-	{
-		return new Param('wf_trigger', 'com.zoho.crm.api.Record.DeleteRecordsParam'); 
-
-	}
-
-}
-
-class GetDeletedRecordsParam
-{
-	public static final function type()
-	{
-		return new Param('type', 'com.zoho.crm.api.Record.GetDeletedRecordsParam'); 
-
-	}
-	public static final function page()
-	{
-		return new Param('page', 'com.zoho.crm.api.Record.GetDeletedRecordsParam'); 
-
-	}
-	public static final function perPage()
-	{
-		return new Param('per_page', 'com.zoho.crm.api.Record.GetDeletedRecordsParam'); 
-
-	}
-
-}
-
-class GetDeletedRecordsHeader
-{
-	public static final function IfModifiedSince()
-	{
-		return new Header('If-Modified-Since', 'com.zoho.crm.api.Record.GetDeletedRecordsHeader'); 
-
-	}
-
-}
-
-class SearchRecordsParam
-{
-	public static final function criteria()
-	{
-		return new Param('criteria', 'com.zoho.crm.api.Record.SearchRecordsParam'); 
-
-	}
-	public static final function email()
-	{
-		return new Param('email', 'com.zoho.crm.api.Record.SearchRecordsParam'); 
-
-	}
-	public static final function phone()
-	{
-		return new Param('phone', 'com.zoho.crm.api.Record.SearchRecordsParam'); 
-
-	}
-	public static final function word()
-	{
-		return new Param('word', 'com.zoho.crm.api.Record.SearchRecordsParam'); 
-
-	}
-	public static final function converted()
-	{
-		return new Param('converted', 'com.zoho.crm.api.Record.SearchRecordsParam'); 
-
-	}
-	public static final function approved()
-	{
-		return new Param('approved', 'com.zoho.crm.api.Record.SearchRecordsParam'); 
-
-	}
-	public static final function page()
-	{
-		return new Param('page', 'com.zoho.crm.api.Record.SearchRecordsParam'); 
-
-	}
-	public static final function perPage()
-	{
-		return new Param('per_page', 'com.zoho.crm.api.Record.SearchRecordsParam'); 
-
-	}
-
-}
-
-class GetMassUpdateStatusParam
-{
-	public static final function jobId()
-	{
-		return new Param('job_id', 'com.zoho.crm.api.Record.GetMassUpdateStatusParam'); 
-
-	}
-
 }
